TASK 6 — “Website Conversation AI” (RAG LLM Chain)/requirements.txt
flask>=2.3
sentence-transformers>=2.2
faiss-cpu>=1.7
llama-cpp-python>=0.1.77
numpy>=1.23
